requirements.txt
fastapi>=0.100.0
langchain>=0.1.0
langchain-openai>=0.0.5
sqlalchemy>=2.0.0
pyodbc>=4.0.39
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=3.4.8
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
uvicorn[standard]>=0.20.0
slowapi>=0.1.9
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0